requirements.txt
wtforms
flask_wtf
flask_login
google-cloud-datastore
google-cloud-storage
google-cloud-logging
google-auth
requests
Flask
slackclient
requests
oauthlib
pyOpenSSL
Flask-Login
flask-slack-events
flask-socketio
mailjet_rest
py-trello
gunicorn
pyyaml